created Contact Preferences section with reducer
