+++ src/state/reducers/index.js
@@ -1,11 +1,11 @@
 import { combineReducers } from 'redux';
 
 import { reducer as client } from './clientReducer';
-import { reducer as documents } from './documentReducer';
+import { reducer as contactPreferences } from './contactPreferencesReducer';
 
 const rootReducer = combineReducers({
   client,
-  documents,
+  contactPreferences,
 });
 
 export default rootReducer;
